chore(dry): abstract pickers into a single component
DataSetPicker, ProgramPicker, UserPicker and TETypePicker components
contained mostly the same code. This code is now inside the generic
ResourcePicker and is being used by various picker inputs.

Consumers use an enum based approach to select what type of
resource they are interested in. The actual queries live in the
`queries.js` file in the same folder as the ResourcePicker.