fix: ee import - refactor to make sure all useEffect dependencies are listed (#1506)
Removed two //eslint-disable comments by refactoring the MappingTable component.

form.mutators don't seem to be memoized, as including them in the
dependencies array was triggering useEffect on every render.
Solution was to use form.change instead, which does seem to be memoized.
This was also actually a cleaner solution as the effect now just sets the whole array
of band-cocs to the array in the earthengine config,
with props id and name instead of bandId and bandName.

Another refactor was to make the components agnostic as to which earth engine
dataset is chosen.