fix: patch webpack to support webworker importScripts with relative publicPath [v39] (#1518)
Fixes https://dhis2.atlassian.net/browse/DHIS2-13850

This patches webpack to fix a bug there causing issues with web workers. It’s not ideal as a generic webpack fix (which is why I haven’t yet opened a PR against webpack). We might want to instead update the publicPath in Create React App to use auto, solving it there instead of in webpack, though that requires more work to check all the edge cases.

However this should be very low-risk as a patch in the import/export app itself, because it only affects web workers and we only use web workers in this one place which has been tested thoroughly and works with the patch in place.