fix: loader keeps showing when opening export in new page
we were assigning setting the state to an event in the new window.
That reference doesn't work one outside the React app in the new window,
so the loader never hides. Not sure if this pattern worked at some point,
but even if it did, it was propbably disabled as a security issue